requirements.txt
python-dotenv
meross-iot
flask